# improve hash efficiency by directly using str/unicode hash